twrp: Disable always rm -rf
Needed for creating partitions. (mkfs.f2fs, etc)

I believe this is safe from the eMMC brick bug. The secure erase command
is removed at the kernel level. So this is another one of the many
redunancies and checks to prevent accidental calls to secure disard.

It is safe to do a mkfs on eMMC brick bug device.