p4-common: Fix Netflix playback
Netflix app looks for a decoder with adaptive-playback feature. The Google
h264 decoder is marked with this feature causing Netflix to pick this
codec but it does not work for this device.

Copy the media_codecs_google_video.xml into device tree and remove the
adaptive-playback feature from the Google h264 decoder.